Added resizeable window functions
Added the `resizeable` and `set_resizeable` functions to check and set whether the window is resizeable.

The original `libui-ng` project opts for the UK spelling "resizeable" as opposed to the US spelling "resizable". I've opted to stick with the former for consistency with the base project, but a spellchecker may mark it as misspelled.